Update readme for Webpack Compatiablity
I think it's important to mention that this plugin is compatible with Webpack 5. As a developer I land on multiple plugins and sometimes this small information is missing and we ended up trying that plugin with the new version and breaking our heads.

So just to be more clear let's mentioned this.